simplify the feature toggle handler further
