+++ magefile.go
@@ -1238,7 +1238,7 @@ func deployFeatureToggles(ctx context.Context, cluster *dev.Cluster) error {
 
 	for _, featTog := range availableFeatureToggles {
 		// feature toggles enabled/disabled at the level of openshift/release in the form of multiple jobs
-		if featTog.IsEnabledOnTestEnv() {
+		if featuretoggle.IsEnabledOnTestEnv(featTog) {
 			if err := featTog.Enable(ctx); err != nil {
 				return fmt.Errorf("failed to enable the feature toggle: %w", err)
 			}
@@ -1256,7 +1256,7 @@ func preClusterCreationFeatureToggleSetup(ctx context.Context) error {
 
 	for _, featTog := range availableFeatureToggles {
 		// feature toggles enabled/disabled at the level of openshift/release in the form of multiple jobs
-		if featTog.IsEnabledOnTestEnv() {
+		if featuretoggle.IsEnabledOnTestEnv(featTog) {
 			if err := featTog.PreClusterCreationSetup(ctx); err != nil {
 				return fmt.Errorf("failed to set the feature toggle before the cluster creation: %w", err)
 			}
@@ -1273,7 +1273,7 @@ func postClusterCreationFeatureToggleSetup(ctx context.Context, cluster *dev.Clu
 
 	for _, featTog := range availableFeatureToggles {
 		// feature toggles enabled/disabled at the level of openshift/release in the form of multiple jobs
-		if featTog.IsEnabledOnTestEnv() {
+		if featuretoggle.IsEnabledOnTestEnv(featTog) {
 			if err := featTog.PostClusterCreationSetup(ctx, cluster); err != nil {
 				return fmt.Errorf("failed to set the feature toggle after the cluster creation: %w", err)
 			}
